added missing reverse operators
